feat: add support for allow re-interaction after a given amount of hours
With the new parameter `can-reinteract-after` your bot is allowed to re-interact someone if a given amount of time has been passed.
For example `can-reinteract-after: 10` allows the bot to re-interact an user that has interacted with more than 10 hours before. You can even specify a fraction of hours (0.5, 0.8, 1.5 or whatever). Ranges are not allowed. Disabled by default.